refactor: friend page의 header도 my page의 header와 같은 마크업 구조로 변경
