+++ services/web/src/app/friend/layout.tsx
@@ -1,9 +1,9 @@
-import { FriendHeaderContent } from '@features/Header';
+import { FriendHeader } from '@features/Header';
 
 function FriendLayout({ children }: { children: React.ReactNode }) {
   return (
     <>
-      <FriendHeaderContent />
+      <FriendHeader />
       {children}
     </>
   )
