--- services/web/src/features/Header/FriendHeaderContent.css.ts
@@ -1,10 +1,0 @@
-import { style } from '@vanilla-extract/css';
-
-export const contentWrapper = style({
-  backgroundColor: 'black',
-});
-
-export const fixedHeader = style({
-  position: 'sticky',
-  top: 0,
-});
