+++ services/web/src/features/Header/index.ts
@@ -1,2 +1,2 @@
 export { default as MyHeader } from './MyHeader';
-export { default as FriendHeaderContent } from './FriendHeaderContent';
+export { default as FriendHeader } from './FriendHeader';
